+++ server/src/handler/eldenring/ghostdata.rs
@@ -36,8 +36,8 @@ impl HandleRequest<Box<RequestCreateGhostDataParams>, ResponseCreateGhostDataPar
         .bind(100)
         .bind(100)
         .bind(&request.replay_data)
-        .bind(request.area.area)
-        .bind(request.area.play_region)
+        .bind(request.area.area as i32)
+        .bind(request.area.play_region as i32)
         .bind(&request.group_passwords)
         .fetch_one(&self.services.database)
         .await?
@@ -59,7 +59,7 @@ impl HandleRequest<Box<RequestGetGhostDataListParams>, ResponseGetGhostDataListP
         let play_regions = request
             .search_areas
             .iter()
-            .map(|a| a.play_region)
+            .map(|a| a.play_region as i32)
             .collect::<Vec<i32>>();
 
         let entries: Vec<ResponseGetGhostDataListParamsEntry> =
@@ -72,8 +72,8 @@ impl HandleRequest<Box<RequestGetGhostDataListParams>, ResponseGetGhostDataListP
             .into_iter()
             .map(|e| ResponseGetGhostDataListParamsEntry {
                 area: PlayRegionArea {
-                    play_region: e.play_region,
-                    area: e.area,
+                    play_region: e.play_region as u32,
+                    area: e.area as u32,
                 },
                 identifier: ObjectIdentifier(e.ghostdata_id),
                 replay_data: e.replay_data,
